Solved a timing problem by switching from 'kpartx -a' to 'kpartx -s -a' Obviously it was returning too fast otherwise,  no partions were showing up in /dev/mapper/

+++ scripts/odroidc1image.sh
@@ -24,23 +24,23 @@ sudo parted -s "${LOOP_DEV}" mkpart primary ext4 65 2113
 sudo parted -s "${LOOP_DEV}" set 1 boot on
 sudo parted -s "${LOOP_DEV}" print
 sudo partprobe "${LOOP_DEV}"
-sudo kpartx -a "${LOOP_DEV}"
- 
+sudo kpartx -s -a "${LOOP_DEV}"
+
 BOOT_PART=`echo /dev/mapper/"$( echo ${LOOP_DEV} | sed -e 's/.*\/\(\w*\)/\1/' )"p1`
 SYS_PART=`echo /dev/mapper/"$( echo ${LOOP_DEV} | sed -e 's/.*\/\(\w*\)/\1/' )"p2`
 echo "Using: " ${BOOT_PART}
 echo "Using: " ${SYS_PART}
-#if [ ! -b "${BOOT_PART}" ]
-#then
-#	echo "${BOOT_PART} doesn't exist"
-#	exit 1
-#fi
-#
-#if [ ! -b "${SYS_PART}" ]
-#then
-#	echo "${SYS_PART} doesn't exist"
-#	exit 1
-#fi
+if [ ! -b "${BOOT_PART}" ]
+then
+	echo "${BOOT_PART} doesn't exist"
+	exit 1
+fi
+
+if [ ! -b "${SYS_PART}" ]
+then
+	echo "${SYS_PART} doesn't exist"
+	exit 1
+fi
 
 echo "Creating filesystems"
 sudo mkfs -t vfat -n BOOT "${BOOT_PART}"
